The actual Stuff
I've got a plan, and even a simple way to do it now,i thought anyways,
but im missing something and cant seem to get it to work. I want to have
a three tab website, one iwth simply a welcome page. the second tab with
a search that searches through a static list of websites and links we
like, and then a third tab that just has our contact info all with
angular. The tabs kinda work, i just havent got stuff into the specific
seperate tags yet, and I haven't filled out welcome or contact yer,
because I've ben so preoccupied trying to get the list to work.